\name{item5fr}
\alias{item5fr}
\title{5-item Test Data Set}
\usage{data(item5fr)}
\description{
A simulated data set with a 5 items, each with 3 ordinal categories labeled 
as 0,1,2. Each row corresponds to a response pattern with the sixth column containing the frequency of each response pattern.
}
\keyword{datasets}

