\name{pln-package}
\alias{pln-package}
\alias{pln}
\docType{package}
\title{
Polytomous Logit-Normit (Graded Logistic) Model Estimation
}
\description{
Performs bivariate composite likelihood and full information maximum
likelihood estimation for polytomous logit-normit (graded logistic) item response theory (IRT) models.
}

\section{Acknowledgment}{
Some code from the ltm package (version 0.9-7) was modified for counting the frequency of response patterns.
}

\details{
\tabular{ll}{
Package: \tab pln\cr
Type: \tab Package\cr
Version: \tab 0.2-2\cr
Date: \tab 2020-07-28\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
This package currently contains several functions performing estimation of unidimensional (single latent
trait) polytomous logit-normit models (also known graded logistic) using bivariate composite likelihood
and full information maximum likelihood estimation.

}
\author{
Carl Falk \email{cffalk@gmail.com},
Harry Joe

Maintainer: Carl F. Falk \email{cffalk@gmail.com}
}
\references{
Bartholomew, D., Knott, M., and Moustaki, I. (2011). \emph{Latent Variable 
Models and Factor Analysis: A Unified Approach}, 3rd Edition. Wiley. 

Maydeu-Olivares, A., and Joe, H. (2005). Limited and full information estimation
and goodness-of-fit testing in \eqn{2^n} contingency tables: A unified framework.
\emph{Journal of the American Statistical Association, 100}, 1009-1020.

Maydeu-Olivares, A., and Joe, H. (2006). Limited information and goodness-of-fit
testing in multidimensional contingency tables. \emph{Psychometrika, 71},
713-732.

Varin, C., Reid, N. and Firth, D. (2011). An overview of composite likelihood
methods. \emph{Statistica Sinica, 21}, 5-42.

}
